pyproject.toml
[tool.poetry]
name = "malcnn"
version = "0.1.0"
description = ""
authors = ["doanhieung <doanhieu.ng@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
scikit-learn = "^1.4.2"
torch = "^2.2.2"
opencv-python = "^4.9.0.80"
tqdm = "^4.66.2"
pillow = "^10.3.0"
pandas = "^2.2.2"
ipykernel = "^6.29.4"
seaborn = "^0.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
